requirements.txt
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.3
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.11.0
Flask==3.0.0
fonttools==4.43.1
fpdf==1.7.2
gitdb==4.0.11
GitPython==3.1.41
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==7.0.1
install==1.3.5
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.8.1
numpy==1.26.1
openai==1.12.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==3.2.0
plotly==5.22.0
prompt_toolkit==3.0.47
protobuf==4.25.2
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-pptx==0.6.22
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
rank-bm25==0.2.2
referencing==0.33.0
regex==2023.10.3
reportlab==4.0.6
requests==2.28.2
rich==13.7.0
rpds-py==0.17.1
scipy==1.12.0
selenium==4.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.31.1
tenacity==8.2.3
tiktoken==0.6.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.22.0
trio-websocket==0.10.2
ttkthemes==3.2.2
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.4.6
urllib3==1.26.15
validators==0.22.0
virtualenv==20.21.0
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.1
Werkzeug==3.0.0
wordcloud==1.9.2
wsproto==1.2.0
XlsxWriter==3.1.9
yarg==0.1.9
zipp==3.17.0
